chore: move the `SSLHTTPAdapter` class to a separate file and its test to the `test` folder
At this point if `requests 2.32.3` is installed, the issue is
reproducable.
