+++ ibm_cloud_sdk_core/utils.py
@@ -19,39 +19,12 @@ import gzip
 import io
 import json as json_import
 import re
-import ssl
 from os import getenv, environ, getcwd
 from os.path import isfile, join, expanduser
 from typing import List, Union
 from urllib.parse import urlparse, parse_qs
 
-from requests.adapters import HTTPAdapter, DEFAULT_POOLBLOCK
-from urllib3.util.ssl_ import create_urllib3_context
-
 import dateutil.parser as date_parser
-
-
-# pylint: disable=fixme
-# TODO: revert the change in the `requirement.txt` once this class become deprecated!
-class SSLHTTPAdapter(HTTPAdapter):
-    """Wraps the original HTTP adapter and adds additional SSL context."""
-
-    def __init__(self, *args, **kwargs):
-        self._disable_ssl_verification = kwargs.pop('_disable_ssl_verification', None)
-
-        super().__init__(*args, **kwargs)
-
-    def init_poolmanager(self, connections, maxsize, block=DEFAULT_POOLBLOCK, **pool_kwargs):
-        """Create and use custom SSL configuration."""
-
-        ssl_context = create_urllib3_context()
-        ssl_context.minimum_version = ssl.TLSVersion.TLSv1_2
-
-        if self._disable_ssl_verification:
-            ssl_context.check_hostname = False
-            ssl_context.verify_mode = ssl.CERT_NONE
-
-        super().init_poolmanager(connections, maxsize, block, ssl_context=ssl_context, **pool_kwargs)
 
 
 class GzipStream(io.RawIOBase):
