+++ test/test_http_adapter.py
@@ -14,6 +14,7 @@ from ibm_cloud_sdk_core.authenticators import NoAuthAuthenticator
 
 
 # The certificate files that are used in this tests are generated by this command:
+# pylint: disable=line-too-long,pointless-string-statement
 """
 openssl req -x509 -out test_ssl.crt -keyout test_ssl.key \
   -newkey rsa:2048 -nodes -sha256 -days 36500 \
@@ -64,7 +65,8 @@ def test_ssl_verification():
         assert res is not None
 
         # Lastly, try with an external URL.
-        # This test case is mainly here to reproduce the regression in the `requests` package that was introduced in `2.32.3`.
+        # This test case is mainly here to reproduce the regression
+        # in the `requests` package that was introduced in `2.32.3`.
         # More details on the issue can be found here: https://github.com/psf/requests/issues/6730
         service = BaseService(service_url='https://raw.githubusercontent.com', authenticator=NoAuthAuthenticator())
         assert service.disable_ssl_verification is False
@@ -77,4 +79,3 @@ def test_ssl_verification():
         server.shutdown()
         # Re-enable warnings.
         warnings.resetwarnings()
-
